+++ g13d/g13_keys.cc
@@ -99,11 +99,13 @@ void G13_Profile::parse_keys(unsigned char *buf) {
 }
 
 G13_Key *G13_Profile::find_key(const std::string &keyname) {
-
   auto key = _keypad.manager().find_g13_key_value(keyname);
-  if (key >= 0 && key < _keys.size()) {
+
+  // TODO(jtgans): Check this is the proper type
+  if (key >= 0 && key < static_cast<int>(_keys.size())) {
     return &_keys[key];
   }
+
   return 0;
 }
